move benchmark tests
